`MarioActor`
Match most of `MarioActor`. `calcAnim` and `calcAndSetBaseMtx` are the functions with the biggest missing match.

+++ src/Game/Util/JointUtil.cpp
@@ -36,7 +36,7 @@ namespace MR {
         return MR::getJ3DModel(pActor)->mMtxBuffer->mpAnmMtx[idx];
     }
 
-    s32 getJointIndex(const LiveActor* pActor, const char* pName) {
+    u16 getJointIndex(const LiveActor* pActor, const char* pName) {
         return MR::getJ3DModelData(pActor)->mJointTree.mJointName->getIndex(pName);
     }
